FIX: random failure on loading comments as soon as they are created
There was an intermittent error (1 out of every 4 times) after the
comment is created is because of comment live reloading (via message
bus) interfering with appending comment after the comment is created.

This commit adds a check before live reloading to ensure that live
reloading does not happen for comment creator.