FIX: make sure QA topic class is applied in all the cases
Earlier the `qa-topic` class was being applied only when the user
entered the topic from the initial few posts. When the user entered from
the bottom of the topic the class was not applied causing styling
issues.

This commit ensures that the QA topic class is applied in all the cases.