DEV: convert post comments widgets to Glimmer components (#149)
Converts all post comments widgets (`post-voting-comments` and all widgets within) to Glimmer components, leaving only `post-voting-post` as a widget. I tried not to differ too much from the original structure to avoid regressions, especially on custom themes.

The main intent of this work is to allow the plugin to still work correctly with browser extensions that manipulate the DOM, like Grammarly, which currently breaks the event handling and makes it unusable.

It also adds a shim to `post-voting-button` to render the Glimmer component when used by the `post-voting-post` widget and to `post-voting-comments` to render the component when decorating `post-menu:after`.