FIX: Post voting topic broken when OP is moderator action
Why this change?

A topic with post voting enabled will fail to render on the client side
when there is only the first post in the topic and the post is converted
to the `moderator_action` subtype.

This plugin adds a custom filter when fetching posts which is supposed to filter out
small actions and whispers. However, we were doing this in an incorrect
manner by filtering only for `regular` posts.

What does this change do?

This change updates our query filter to filter way whispers and small
actions instead of just filtering for regular topics.