DEV: Fix mismatched column types
The primary key is usually a bigint column, but the foreign key columns
are usually of integer type. This can lead to issues when joining these
columns due to mismatched types and different value ranges.

This was using a temporary plugin / test API to make tests pass, but
it is safe to alter these tables because they usually have less than 1M
rows.